CreateExercise Unit Tests [surely not all] -Msi
