+++ nest-back/test/Workout/UnitTests/createExerciseUnit.spec.ts
@@ -5,7 +5,7 @@ import {v4 as uuidv4 } from 'uuid';
 import {
     Workout,
     Exercise,
-    Planner,
+    User,
     Difficulty,
     Prisma
 } from '@prisma/client';
@@ -32,7 +32,6 @@ test('should create new exercise ', async () => {
         Posedescription:"TestPDesc",
         restPeriod:2,
         difficulty:Difficulty.EASY,
-
         duratime:2,
         workoutWorkoutID: null
     }
@@ -40,5 +39,145 @@ test('should create new exercise ', async () => {
 
     await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).resolves.toEqual(
         "Exercise created."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Title)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"TestRange",
+        sets:4,
+        Posedescription:"TestPDesc",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Description)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"TestTitle",
+        description:"",
+        repRange:"TestRange",
+        sets:4,
+        Posedescription:"TestPDesc",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Rep range)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"",
+        sets:4,
+        Posedescription:"TestPDesc",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Sets)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"",
+        sets:0,
+        Posedescription:"TestPDesc",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Pose Description)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"",
+        sets:4,
+        Posedescription:"",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Rest Period)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"",
+        sets:4,
+        Posedescription:"",
+        restPeriod:0,
+        difficulty:Difficulty.EASY,
+        duratime:2,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
+    )
+})
+
+test('Should not create workout [Throws Empty Parameters error(Duration)]  ', async () => {
+    const Exercise = {
+        exercise:uuidv4(),
+        title:"",
+        description:"TestDescription",
+        repRange:"",
+        sets:4,
+        Posedescription:"",
+        restPeriod:2,
+        difficulty:Difficulty.EASY,
+        duratime:0,
+        workoutWorkoutID: null
+    }
+    mockCtx.prisma.exercise.create.mockResolvedValue(Exercise)
+
+    await expect(workoutService.createExercise(Exercise.title,Exercise.description,Exercise.repRange,Exercise.sets,Exercise.Posedescription,Exercise.restPeriod,Exercise.difficulty,Exercise.duratime,ctx)).rejects.toThrow(
+        "Parameters can not be left empty."
     )
 })
